jobbright/backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
passlib[bcrypt]
pydantic[email]
sqlalchemy
psycopg2-binary
alembic
celery[redis]
python-jose[cryptography]
requests
beautifulsoup4
playwright
qdrant-client
sentence-transformers
numpy
stripe
PyPDF2 